refactor: co-locate templates with create command
