+++ lib/src/commands/create/create.dart
@@ -7,7 +7,7 @@ import 'package:universal_io/io.dart';
 import 'package:usage/usage_io.dart';
 import 'package:very_good_analysis/very_good_analysis.dart';
 import 'package:very_good_cli/src/command_runner.dart';
-import 'package:very_good_cli/src/templates/templates.dart';
+import 'package:very_good_cli/src/commands/create/templates/templates.dart';
 
 const _defaultOrgName = 'com.example.verygoodcore';
 const _defaultDescription = 'A Very Good Project created by Very Good CLI.';
