+++ lib/src/commands/create/templates/dart_package/dart_package.dart
