+++ lib/src/commands/create/templates/dart_package/dart_package_bundle.dart
