+++ lib/src/commands/create/templates/flutter_package/flutter_package.dart
