+++ lib/src/commands/create/templates/flutter_package/flutter_package_bundle.dart
