+++ lib/src/commands/create/templates/flutter_plugin/flutter_plugin.dart
