+++ lib/src/commands/create/templates/post_generate_actions.dart
