+++ lib/src/commands/create/templates/very_good_core/very_good_core.dart
