+++ lib/src/commands/create/templates/very_good_core/very_good_core_template.dart
@@ -1,6 +1,6 @@
 import 'package:mason/mason.dart';
 import 'package:universal_io/io.dart';
-import 'package:very_good_cli/src/templates/templates.dart';
+import 'package:very_good_cli/src/commands/create/templates/templates.dart';
 
 /// {@template very_good_core_template}
 /// A core Flutter app template.
